Attach new AI
