+++ Assets/Team4/Scripts/Players/AiBrain.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace MatrixJam.Team4
+{
+    public abstract class AiBrain : MonoBehaviour
+    {
+        public abstract TurnObject MakeDecision(TurnData turnData, Player player);
+    }
+}
