--- Assets/Team4/Scripts/Players/FirstOptionPlayer.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c12ee0e8479040ff987f52bcb51253b0
-timeCreated: 1596903162
